--- mu4/scores/qml/MuseScore/Scores/qmldir
@@ -1,4 +1,0 @@
-module MuseScore.Scores
-ScoresModule 1.0 ScoresModule.qml
-NewScoreDialog 1.0 NewScoreDialog.qml
-ScoreItem 1.0 ScoreItem.qml
