--- mu4/scores/scores.qrc
@@ -1,8 +1,0 @@
-<RCC>
-    <qresource prefix="/">
-        <file>qml/MuseScore/Scores/ScoresModule.qml</file>
-        <file>qml/MuseScore/Scores/qmldir</file>
-        <file>qml/MuseScore/Scores/NewScoreDialog.qml</file>
-        <file>qml/MuseScore/Scores/ScoreItem.qml</file>
-    </qresource>
-</RCC>
