+++ mu4/userscores/qml/MuseScore/UserScores/qmldir
@@ -1,0 +1,4 @@
+module MuseScore.UserScores
+UserScoresModule 1.0 UserScoresModule.qml
+NewScoreDialog 1.0 NewScoreDialog.qml
+ScoreItem 1.0 ScoreItem.qml
